pyproject.toml
[project]
name = "rossa"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "yfinance>=1.2.2",
    "matplotlib>=3.8.0",
    "networkx>=3.2",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.2.0",
    "scipy>=1.17.1",
    "statsmodels>=0.14.6",
    "streamlit>=1.44.0",
    "tqdm>=4.67.3",
    "n100tickers",
    "edgartools>=5.30.0",
]

[tool.uv.workspace]
members = [
    "n100tickers",
]

[tool.uv.sources]
n100tickers = { workspace = true, editable = true }
